+++ app/specific/Main.js
@@ -24,8 +24,8 @@ var Main_isTV;
 var Main_isDebug = false;
 
 var Main_stringVersion = '3.0';
-var Main_stringVersion_Min = '.268';
-var Main_version_java = 268;//Always update (+1 to current value) Main_version_java after update Main_stringVersion_Min or a major update of the apk is released
+var Main_stringVersion_Min = '.269';
+var Main_version_java = 269;//Always update (+1 to current value) Main_version_java after update Main_stringVersion_Min or a major update of the apk is released
 var Main_minversion = 'October 22 2020';
 var Main_version_web = 506;//Always update (+1 to current value) Main_version_web after update Main_minversion or a major update of the web part of the app
 var Main_versionTag = Main_stringVersion + Main_stringVersion_Min + '-' + Main_minversion;
@@ -617,8 +617,9 @@ function Main_Changelog() {
 
     var changelogObj = [
         {
-            title: "Apk Version 3.0.268 - Web Version October 22 2020",
+            title: "Apk Version 3.0.268 and 3.0.269 - Web Version October 22 2020",
             changes: [
+                "Fix preview volume control",
                 "Automatic cleanup history unreachable Lives that became a VOD but the VOD is now deleted",
                 "General performance improves and bug fixes",
             ]