+++ apk/Changelog.md
@@ -9,9 +9,10 @@ The App is divided in two part, **Apk** and **Web** you can check it version in 
 
 The app has a building update check function, so whenever there is a update a red text will show up bellow the main pages clock, also a short warning will pop on the screen let you know if is a **Apk** or **Web** update.
 
-Apk Version 3.0.268 - Web Version October 22 2020
+Apk Version 3.0.268 and 3.0.269 - Web Version October 22 2020
 ==============
 
+* Fix preview volume control
 * Automatic cleanup history unreachable Lives that became a VOD but the VOD is now deleted
 * General performance improves and bug fixes